Refactor: Move obsidian markdown file class to own file
